#43. context_render_backend_system: Use lower integers in places where there is no practical sense to use 64 bits.

+++ unit/context_render_backend_system_vulkan/kan/context/buffer.c
@@ -3,7 +3,7 @@
 struct render_backend_buffer_t *render_backend_system_create_buffer (struct render_backend_system_t *system,
                                                                      enum render_backend_buffer_family_t family,
                                                                      enum kan_render_buffer_type_t buffer_type,
-                                                                     uint64_t full_size,
+                                                                     uint32_t full_size,
                                                                      kan_interned_string_t tracking_name)
 {
     VkBufferUsageFlags usage_flags = 0u;
@@ -181,7 +181,7 @@ void render_backend_system_destroy_buffer (struct render_backend_system_t *syste
 
 kan_render_buffer_t kan_render_buffer_create (kan_render_context_t context,
                                               enum kan_render_buffer_type_t type,
-                                              uint64_t full_size,
+                                              uint32_t full_size,
                                               void *optional_initial_data,
                                               kan_interned_string_t tracking_name)
 {
@@ -238,7 +238,7 @@ kan_render_buffer_t kan_render_buffer_create (kan_render_context_t context,
     return (kan_render_buffer_t) buffer;
 }
 
-void *kan_render_buffer_patch (kan_render_buffer_t buffer, uint64_t slice_offset, uint64_t slice_size)
+void *kan_render_buffer_patch (kan_render_buffer_t buffer, uint32_t slice_offset, uint32_t slice_size)
 {
     struct render_backend_buffer_t *data = (struct render_backend_buffer_t *) buffer;
     struct render_backend_schedule_state_t *schedule = render_backend_system_get_schedule_for_memory (data->system);
